consumption_framework-main/requirements.txt
click>=8.1.3
tqdm>=4.66.1
elasticsearch>=8.3.3
pandas>=1.5.0
numpy>=1.23.2
PyYAML>=6.0.1
requests>=2.31.0
pytz>=2023.3.post1
